refactor with added dependency to lib module
