+++ lib/build.gradle
@@ -23,5 +23,6 @@ dependencies {
     implementation "org.slf4j:slf4j-simple:$slf4j_simple_ver"
     implementation("com.github.ajalt.clikt:clikt:3.0.0")
 
+    testImplementation group: 'org.mockito', name: 'mockito-core', version: '3.3.3'
     testImplementation group: 'junit', name: 'junit', version: '4.12'
 }
