+++ lib/src/main/kotlin/org/wycliffeassociates/resourcecontainer/media/BookMediaDownloader.kt
@@ -4,16 +4,18 @@ import java.io.File
 import org.wycliffeassociates.resourcecontainer.media.data.MediaDivision
 import org.wycliffeassociates.resourcecontainer.media.data.MediaUrlParameter
 import org.wycliffeassociates.resourcecontainer.media.io.DownloadClient
+import org.wycliffeassociates.resourcecontainer.media.io.IDownloadClient
 
 class BookMediaDownloader(
     rcFile: File,
     overwrite: Boolean,
-    urlParams: MediaUrlParameter
-) : RCMediaDownloader(rcFile, overwrite, urlParams) {
+    urlParams: MediaUrlParameter,
+    downloadClient: IDownloadClient
+) : RCMediaDownloader(rcFile, overwrite, urlParams, downloadClient) {
 
     override fun downloadMedia(url: String): String {
         val contentDir = createTempDir()
-        val downloadedFile = DownloadClient.downloadFromUrl(url, contentDir)
+        val downloadedFile = downloadClient.downloadFromUrl(url, contentDir)
 
         if (downloadedFile != null) {
             val pathInRC = "$MEDIA_DIR/${urlParams.projectId}/${downloadedFile.name}"
