+++ lib/src/main/kotlin/org/wycliffeassociates/resourcecontainer/media/io/IDownloadClient.kt
@@ -1,0 +1,7 @@
+package org.wycliffeassociates.resourcecontainer.media.io
+
+import java.io.File
+
+interface IDownloadClient {
+    fun downloadFromUrl(url: String, outputDir: File): File?
+}
